Implement app factory design
# Conflicts:
#	linnote/__main__.py
