+++ linnote/__main__.py
@@ -1,14 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-u"""
-Main script for linnote.
+from linnote.client import create_app
 
-Author: Anatole Hanniet, Tutorat Santé Lyon Sud (2014-2017).
-License: Mozilla Public License, see 'LICENSE.txt' for details.
-"""
-
-from linnote.client import APP
-
-
-APP.run(host='0.0.0.0')
+app = create_app('linnote')
+app.run()
